Styles and component refactor
