+++ plugin-hrm-form/src/components/case/casePrint/mockedData.ts
@@ -1,0 +1,192 @@
+export const callerInfoSection = {
+  sectionName: 'Caller Information',
+  fieldValues: [
+    { label: 'First Name', value: 'Ross' },
+    { label: 'Last Name', value: 'Keller' },
+    { label: 'Address', value: '------' },
+    { label: 'Province', value: '------' },
+    { label: 'Municipality', value: '------' },
+    { label: 'District', value: '------' },
+    { label: 'Contact Number', value: '031 201 1111' },
+    { label: 'Relationship to Child', value: 'Teacher' },
+    { label: 'Gender', value: 'Boy' },
+    { label: 'Age', value: '12' },
+    { label: 'Language', value: 'English' },
+    { label: 'Race', value: 'Indian' },
+  ],
+};
+
+export const childInfoSection = {
+  sectionName: 'Child Information',
+  fieldValues: [
+    { label: 'First Name', value: 'Child' },
+    { label: 'Last Name', value: 'ZA' },
+    { label: 'Address', value: '------' },
+    { label: 'Province', value: '------' },
+    { label: 'Municipality', value: '------' },
+    { label: 'District', value: '------' },
+    { label: 'Contact Number', value: '------' },
+    { label: 'Gender', value: 'Boy' },
+    { label: 'Age', value: '12' },
+    { label: 'Language', value: 'English' },
+    { label: 'Race', value: 'Indian' },
+    { label: 'School Name', value: '------' },
+    { label: 'School Address', value: '------' },
+    { label: 'Educator', value: '------' },
+    { label: 'Grade Level', value: '8' },
+    { label: 'Living Situation', value: 'Unknown' },
+    { label: 'Region', value: '------' },
+    { label: 'Child HIV Positive', value: '------' },
+    { label: 'Child in conflict with the law', value: '------' },
+    { label: 'Child in detention', value: '------' },
+    { label: 'Child member of an ethnic or racial minority', value: 'No' },
+    { label: 'Child on the move or in migration', value: '------' },
+    { label: 'Child with disability', value: 'Yes' },
+    { label: 'LGBTQI+ / SOEGIS Child', value: 'Unknown' },
+  ],
+};
+
+export const contactSection = {
+  sectionName: 'Contact',
+  fieldValues: [
+    {
+      label: 'Contact Summary',
+      value:
+        'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam convallis magna a suscipit scelerisque. Vestibulum molestie mi ipsum, eu elementum augue malesuada at.',
+    },
+    { label: 'Channel', value: 'SMS' },
+    { label: 'Phone Number', value: '+12066127815' },
+    { label: 'Converstion Duration', value: '21m 30s' },
+    { label: 'Counsellor', value: 'Jana Kleitsch' },
+    { label: 'Date/Time', value: 'Feb 3 2021, 10:30 pm' },
+    { label: 'Repeat Caller?', value: 'Unknown' },
+    { label: 'Referred To', value: 'Unknown' },
+    { label: 'How did the child hear about us?', value: 'Unknown' },
+    { label: 'Keep confidential?', value: 'Yes' },
+    { label: 'OK for case worker to call?', value: 'Unknown' },
+    { label: 'Did you discuss right with the child?', value: 'Unknown' },
+    { label: 'Did the child feel we solved their problem?', value: 'Unknown' },
+    { label: 'Would the child recommend us to a friend?', value: 'Unknown' },
+  ],
+};
+
+export const incidentSection = {
+  sectionName: 'Incidents',
+  fieldValues: [
+    { label: 'Date', value: '11/03/2020' },
+    { label: 'Duration', value: 'Ongoing' },
+    { label: 'Location', value: 'Home' },
+    { label: 'Is caregiver aware?', value: 'No' },
+    { label: 'Was this incident witnessed by anyone', value: 'No' },
+  ],
+};
+
+export const householdMultiSection = {
+  sectionName: 'Household Members',
+  sectionValues: [
+    {
+      key: 1,
+      fieldValues: [
+        { label: 'First Name', value: 'John' },
+        { label: 'Last Name', value: 'McKinley' },
+        { label: 'Address', value: '------' },
+        { label: 'Province', value: '------' },
+        { label: 'Municipality', value: '------' },
+        { label: 'District', value: '------' },
+        { label: 'Contact Number', value: '031 201 1111' },
+        { label: 'Relationship to Child', value: 'Parent' },
+        { label: 'Gender', value: 'Boy' },
+        { label: 'Age', value: '>25' },
+        { label: 'Language', value: 'English' },
+      ],
+    },
+    {
+      key: 2,
+      fieldValues: [
+        { label: 'First Name', value: 'Janice' },
+        { label: 'Last Name', value: 'McKinley' },
+        { label: 'Address', value: '------' },
+        { label: 'Province', value: '------' },
+        { label: 'Municipality', value: '------' },
+        { label: 'District', value: '------' },
+        { label: 'Contact Number', value: '031 201 1111' },
+        { label: 'Relationship to Child', value: 'Parent' },
+        { label: 'Gender', value: 'Girl' },
+        { label: 'Age', value: '>25' },
+        { label: 'Language', value: 'English' },
+      ],
+    },
+  ],
+};
+
+export const perpetratorMultiSection = {
+  sectionName: 'Perpetrator',
+  sectionValues: [
+    {
+      key: 1,
+      fieldValues: [
+        { label: 'First Name', value: 'Janice' },
+        { label: 'Last Name', value: 'McKinley' },
+        { label: 'Address', value: '------' },
+        { label: 'Province', value: '------' },
+        { label: 'Municipality', value: '------' },
+        { label: 'District', value: '------' },
+        { label: 'Contact Number', value: '031 201 1111' },
+        { label: 'Relationship to Child', value: 'Neighbor' },
+        { label: 'Gender', value: 'Girl' },
+        { label: 'Age', value: '>25' },
+        { label: 'Language', value: 'English' },
+      ],
+    },
+    {
+      key: 2,
+      fieldValues: [
+        { label: 'First Name', value: 'Unknown' },
+        { label: 'Last Name', value: 'Unknown' },
+        { label: 'Address', value: '------' },
+        { label: 'Province', value: '------' },
+        { label: 'Municipality', value: '------' },
+        { label: 'District', value: '------' },
+        { label: 'Contact Number', value: '031 201 1111' },
+        { label: 'Relationship to Child', value: 'Neighbor' },
+        { label: 'Gender', value: 'Girl' },
+        { label: 'Age', value: '>25' },
+        { label: 'Language', value: 'English' },
+      ],
+    },
+  ],
+};
+
+export const referralsSection = {
+  sectionName: 'Referrals',
+  fieldValues: [
+    { label: 'Added by', value: 'John Doe' },
+    { label: 'Date', value: '1/3/2021' },
+    { label: 'Recommendation', value: 'State Agency 1' },
+    { label: 'Comments', value: 'Lorem ipsum dolor sit amet' },
+  ],
+};
+
+export const notesSection = {
+  notes: [
+    {
+      counselor: 'Some Counselor',
+      date: '11/03/2020',
+      note:
+        'Sed tincidunt odio eget nisi semper euismod. Vestibulum commodo vulputate sem, vel finibus augue. Proin ultricies faucibus urna, cursus consequat augue volutpat at. Mauris in tellus neque. Sed finibus rhoncus odio bibendum pellentesque. Etiam quam ex, ultrices nec efficitur eu, semper ut nulla. Maecenas libero arcu, molestie et finibus quis, efficitur ut turpis. Fusce iaculis mollis rutrum.',
+    },
+    {
+      counselor: 'Some Counselor',
+      date: '15/03/2020',
+      note:
+        'Proin arcu felis, rhoncus eget pretium id, porta sit amet elit. Quisque sed sem convallis mi vehicula tristique. Cras laoreet cursus odio, vitae vestibulum odio blandit lacinia.',
+    },
+  ],
+};
+
+export const summary =
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam convallis magna a suscipit scelerisque. Vestibulum molestie mi ipsum, eu elementum augue malesuada at. Phasellus ac nibh in nibh ullamcorper luctus. Fusce tristique dui nulla. In lacinia sem a mi congue, eget tincidunt risus euismod. Aliquam euismod metus eu augue ultrices, non sagittis tellus aliquam. Phasellus non lacus et augue convallis lacinia et ac nisl. Nulla interdum lectus eget nulla pulvinar pretium. Aenean laoreet enim vitae diam tristique, dapibus suscipit ligula placerat. Maecenas pellentesque egestas metus sit amet ornare. Proin vel dui nulla. Nam fringilla venenatis justo in porta.';
+
+export const caseManager = { name: 'Bhavna Lutchman', phone: '031 201 2059', email: 'research@childlinesa.org.za' };
+
+export const officeName = 'Gauteng';
