+++ plugin-hrm-form/src/components/case/casePrint/styles/notes.ts
@@ -1,0 +1,12 @@
+import ReactPDF from '@react-pdf/renderer';
+
+const notesStyles: ReactPDF.Styles = {
+  noteSummaryText: {
+    marginLeft: 10,
+    marginRight: 10,
+    marginTop: 10,
+    marginBottom: 10,
+  },
+};
+
+export default notesStyles;
