+++ plugin-hrm-form/src/components/case/casePrint/styles/summary.ts
@@ -1,0 +1,10 @@
+import ReactPDF from '@react-pdf/renderer';
+
+const summaryStyles: ReactPDF.Styles = {
+  caseSummaryText: {
+    fontFamily: 'Open Sans',
+    fontSize: 12,
+  },
+};
+
+export default summaryStyles;
